my lib

+++ Common_Core/my_libft/print/ft_putnbr_base.c
@@ -12,7 +12,7 @@
 
 #include "../includes/libft.h"
 
-int	ft_putnbr_base(long unsigned int nbr, unsigned int base, int is_upper)
+int	ft_putnbr_base(long long int nbr, unsigned int base, int is_upper)
 {
 	int		digits;
 	char	*base_chars;
